services/notifier/pyproject.toml
[project]
name = "notifier"
version = "0.1.0"
description = "Observation alerting and system heartbeat monitoring service"
authors = [
    {name = "Nikos Zacharatos", email = "nikos.zacharatos@iccs.gr"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "redis (>=7.3.0,<8.0.0)",
    "aioredis (>=2.0.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0,<7.0)",
    "loguru (>=0.7.2,<1.0.0)",
    "docker (>=7.0.0,<8.0.0)",
    "shared @ file:///home/nikos/home-telemetry/shared",
]

[project.optional-dependencies]
test = [
    "pytest (>=7.4.0,<8.0.0)",
    "pytest-asyncio (>=0.23.0,<1.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
]

[tool.poetry.dependencies]
pytest = { version = ">=7.4.0,<8.0.0", optional = true }
pytest-asyncio = { version = ">=0.23.0,<1.0.0", optional = true }
pytest-cov = { version = ">=4.1.0,<5.0.0", optional = true }

[tool.poetry.extras]
test = ["pytest", "pytest-asyncio", "pytest-cov"]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
